@@ -12,41 +13,1364 @@
     "name": "TJ Holowaychuk",
     "email": "tj@vision-media.ca"
   },
+  "contributors": [
+    {
+      "name": "aaroncrows",
+      "url": "https://github.com/aaroncrows"
+    },
+    {
+      "name": "Aaron Hamid",
+      "email": "aaron@incsw.com",
+      "url": "https://github.com/ahamid"
+    },
+    {
+      "name": "Aaron Heckmann",
+      "email": "aaron.heckmann+github@gmail.com",
+      "url": "https://github.com/aheckmann"
+    },
+    {
+      "name": "Adam Crabtree",
+      "url": "CrabDude's alias"
+    },
+    {
+      "name": "Adam Gruber",
+      "url": "https://github.com/adamgruber"
+    },
+    {
+      "name": "Adrian Ludwig",
+      "url": "https://github.com/adrian-ludwig"
+    },
+    {
+      "name": "Ainthe Kitchen",
+      "email": "a.in.the.k@gmail.com",
+      "url": "https://github.com/ainthek"
+    },
+    {
+      "name": "ajaykodali",
+      "url": "https://github.com/ajaykodali"
+    },
+    {
+      "name": "Alex Early",
+      "url": "https://github.com/aearly"
+    },
+    {
+      "name": "Alex Pham",
+      "email": "thedark1337@thedark1337.com",
+      "url": "https://github.com/thedark1337"
+    },
+    {
+      "name": "amsul",
+      "url": "https://github.com/amsul"
+    },
+    {
+      "name": "Andreas Brekken",
+      "email": "andreas@brekken.com",
+      "url": "https://github.com/abrkn"
+    },
+    {
+      "name": "Andreas Lind",
+      "email": "andreas@one.com",
+      "url": "https://github.com/papandreou"
+    },
+    {
+      "name": "Andrew Miller",
+      "email": "vnikitin@live.com",
+      "url": "https://github.com/vnikiti"
+    },
+    {
+      "name": "Andrew Nesbitt",
+      "email": "andrewnez@gmail.com",
+      "url": "https://github.com/andrew"
+    },
+    {
+      "name": "Andrey Popp",
+      "email": "8mayday@gmail.com",
+      "url": "https://github.com/andreypopp"
+    },
+    {
+      "name": "Andrii Shumada",
+      "email": "eagleeyes91@gmail.com",
+      "url": "https://github.com/eagleeye"
+    },
+    {
+      "name": "Anis Safine",
+      "url": "https://github.com/anis"
+    },
+    {
+      "name": "Arian Stolwijk",
+      "email": "stolwijk.arian@gmail.com",
+      "url": "https://github.com/arian"
+    },
+    {
+      "name": "Ariel Mashraki",
+      "email": "ariel@mashraki.co.il",
+      "url": "https://github.com/a8m"
+    },
+    {
+      "name": "Arnaud Brousseau",
+      "url": "https://github.com/ArnaudBrousseau"
+    },
+    {
+      "name": "Atsuya Takagi",
+      "email": "atsuya.takagi@gmail.com",
+      "url": "https://github.com/atsuya"
+    },
+    {
+      "name": "Attila Domokos",
+      "url": "https://github.com/adomokos"
+    },
+    {
+      "name": "Austin Birch",
+      "url": "https://github.com/austinbirch"
+    },
+    {
+      "name": "Avi Vahl",
+      "url": "https://github.com/AviVahl"
+    },
+    {
+      "name": "Ben Bradley",
+      "url": "https://github.com/ben-bradley"
+    },
+    {
+      "name": "beneidel",
+      "url": "https://github.com/beneidel"
+    },
+    {
+      "name": "Benjie Gillam",
+      "url": "https://github.com/benjie"
+    },
+    {
+      "name": "Ben Noordhuis",
+      "email": "info@bnoordhuis.nl",
+      "url": "https://github.com/bnoordhuis"
+    },
+    {
+      "name": "Benoit Larroque",
+      "url": "https://github.com/zetaben"
+    },
+    {
+      "name": "Benoît Zugmeyer",
+      "url": "https://github.com/BenoitZugmeyer"
+    },
+    {
+      "name": "Ben Vinegar",
+      "url": "https://github.com/benvinegar"
+    },
+    {
+      "name": "Berker Peksag",
+      "email": "berker.peksag@gmail.com",
+      "url": "https://github.com/berkerpeksag"
+    },
+    {
+      "name": "Bjørge Næss",
+      "url": "https://github.com/bjoerge"
+    },
+    {
+      "name": "Brendan Nee",
+      "email": "brendan@blinktag.com",
+      "url": "https://github.com/brendannee"
+    },
+    {
+      "name": "Brian Beck",
+      "email": "exogen@gmail.com",
+      "url": "https://github.com/exogen"
+    },
+    {
+      "name": "Brian C",
+      "email": "brian.m.carlson@gmail.com",
+      "url": "https://github.com/brianc"
+    },
+    {
+      "name": "Brian Lalor",
+      "email": "blalor@bravo5.org",
+      "url": "https://github.com/blalor"
+    },
+    {
+      "name": "Brian Moore",
+      "url": "https://github.com/bionicbrian"
+    },
+    {
+      "name": "Bryan Donovan",
+      "url": "https://github.com/BryanDonovan"
+    },
+    {
+      "name": "Buck Doyle",
+      "url": "https://github.com/backspace"
+    },
+    {
+      "name": "C. Scott Ananian",
+      "email": "cscott@cscott.net",
+      "url": "https://github.com/cscott"
+    },
+    {
+      "name": "Casey Foster",
+      "url": "https://github.com/caseywebdev"
+    },
+    {
+      "name": "Charles Lowell",
+      "email": "cowboyd@frontside.io",
+      "url": "https://github.com/cowboyd"
+    },
+    {
+      "name": "Chris Buckley",
+      "email": "chris@cmbuckley.co.uk",
+      "url": "https://github.com/cmbuckley"
+    },
+    {
+      "name": "Christopher Hiller",
+      "email": "boneskull@boneskull.com",
+      "url": "https://github.com/boneskull"
+    },
+    {
+      "name": "Chris Wren",
+      "email": "chriswrendev@gmail.com",
+      "url": "https://github.com/ChrisWren"
+    },
+    {
+      "name": "Clemens Stolle",
+      "url": "https://github.com/klaemo"
+    },
+    {
+      "name": "Connor Dunn",
+      "url": "https://github.com/Connorhd"
+    },
+    {
+      "name": "Corey Butler",
+      "url": "https://github.com/coreybutler"
+    },
+    {
+      "name": "Cory Thomas",
+      "url": "https://github.com/dump247"
+    },
+    {
+      "name": "cybertk",
+      "url": "https://github.com/cybertk"
+    },
+    {
+      "name": "Daniel Ericsson",
+      "url": "https://github.com/monowerker"
+    },
+    {
+      "name": "Daniel St. Jules",
+      "email": "danielst.jules@gmail.com",
+      "url": "https://github.com/danielstjules"
+    },
+    {
+      "name": "Daniel Stockman",
+      "email": "daniel.stockman@gmail.com",
+      "url": "https://github.com/evocateur"
+    },
+    {
+      "name": "Dave McKenna",
+      "email": "davemckenna01@gmail.com",
+      "url": "https://github.com/davemckenna01"
+    },
+    {
+      "name": "David da Silva",
+      "email": "yo@dasilvacont.in",
+      "url": "https://github.com/dasilvacontin"
+    },
+    {
+      "name": "David Henderson",
+      "url": "https://github.com/dhendo"
+    },
+    {
+      "name": "Denis Bardadym",
+      "email": "bardadymchik@gmail.com",
+      "url": "https://github.com/btd"
+    },
+    {
+      "name": "Devin Weaver",
+      "email": "suki@tritarget.org",
+      "url": "https://github.com/sukima"
+    },
+    {
+      "name": "Diogo Monteiro",
+      "email": "diogo.gmt@gmail.com",
+      "url": "https://github.com/diogogmt"
+    },
+    {
+      "name": "Dmitry Shirokov",
+      "email": "deadrunk@gmail.com",
+      "url": "https://github.com/runk"
+    },
+    {
+      "name": "Domenic Denicola",
+      "email": "d@domenic.me",
+      "url": "https://github.com/domenic"
+    },
+    {
+      "name": "Dominic Barnes",
+      "email": "dominic@dbarnes.info",
+      "url": "https://github.com/dominicbarnes"
+    },
+    {
+      "name": "domq",
+      "url": "https://github.com/domq"
+    },
+    {
+      "name": "Douglas Wilson",
+      "email": "doug@somethingdoug.com",
+      "url": "https://github.com/dougwilson"
+    },
+    {
+      "name": "Duncan Beevers",
+      "email": "duncan@dweebd.com",
+      "url": "https://github.com/duncanbeevers"
+    },
+    {
+      "name": "Duncan Wong",
+      "url": "https://github.com/badunk"
+    },
+    {
+      "name": "eiji.ienaga",
+      "url": "https://github.com/haru01"
+    },
+    {
+      "name": "Fabio Crisci",
+      "email": "piuccio@gmail.com",
+      "url": "https://github.com/piuccio"
+    },
+    {
+      "name": "Fede Ramirez",
+      "email": "i@2fd.me",
+      "url": "https://github.com/2fd"
+    },
+    {
+      "name": "Fedor Indutny",
+      "email": "fedor@indutny.com",
+      "url": "https://github.com/indutny"
+    },
+    {
+      "name": "fengmk2",
+      "email": "m@fengmk2.com",
+      "url": "https://github.com/fengmk2"
+    },
+    {
+      "name": "Florian Margaine",
+      "email": "florian@margaine.com",
+      "url": "https://github.com/ralt"
+    },
+    {
+      "name": "Forbes Lindesay",
+      "url": "https://github.com/ForbesLindesay"
+    },
+    {
+      "name": "Frederico Silva",
+      "url": "https://github.com/fredericosilva"
+    },
+    {
+      "name": "Fredrik Enestad",
+      "email": "fredrik.enestad@soundtrackyourbrand.com",
+      "url": "https://github.com/fredr"
+    },
+    {
+      "name": "Fredrik Lindin",
+      "url": "https://github.com/Cowboy-coder"
+    },
+    {
+      "name": "Gabriel Silk",
+      "url": "https://github.com/gsilk"
+    },
+    {
+      "name": "Gareth Aye",
+      "email": "gareth.aye@gmail.com",
+      "url": "https://github.com/gaye"
+    },
+    {
+      "name": "Gavin Mogan",
+      "email": "gavin@gavinmogan.com",
+      "url": "https://github.com/halkeye"
+    },
+    {
+      "name": "gigadude",
+      "url": "https://github.com/gigadude"
+    },
+    {
+      "name": "Giovanni Bassi",
+      "url": "https://github.com/giggio"
+    },
+    {
+      "name": "Glen Huang",
+      "email": "curvedmark@gmail.com",
+      "url": "https://github.com/curvedmark"
+    },
+    {
+      "name": "Glen Mailer",
+      "email": "glen@stainlessed.co.uk",
+      "url": "https://github.com/glenjamin"
+    },
+    {
+      "name": "Greg Perkins",
+      "email": "gregperkins@alum.mit.edu",
+      "url": "https://github.com/gregrperkins"
+    },
+    {
+      "name": "Guillermo Rauch",
+      "email": "rauchg@gmail.com",
+      "url": "https://github.com/rauchg"
+    },
+    {
+      "name": "Guy Arye",
+      "url": "https://github.com/aryeguy"
+    },
+    {
+      "name": "Gyandeep Singh",
+      "email": "gyandeeps@gmail.com",
+      "url": "https://github.com/gyandeeps"
+    },
+    {
+      "name": "Harish",
+      "email": "hyeluri@gmail.com",
+      "url": "https://github.com/hyeluri"
+    },
+    {
+      "name": "Harry Brundage",
+      "email": "harry.brundage@gmail.com",
+      "url": "https://github.com/airhorns"
+    },
+    {
+      "name": "Ian Remmel",
+      "email": "design@ianwremmel.com",
+      "url": "https://github.com/ianwremmel"
+    },
+    {
+      "name": "Ian Storm Taylor",
+      "url": "https://github.com/ianstormtaylor"
+    },
+    {
+      "name": "Ian Young",
+      "email": "ian@iangreenleaf.com",
+      "url": "https://github.com/iangreenleaf"
+    },
+    {
+      "name": "Ivan",
+      "url": "https://github.com/ivanstoyanov"
+    },
+    {
+      "name": "Jaakko Salonen",
+      "url": "https://github.com/jsalonen"
+    },
+    {
+      "name": "Jacob Wejendorp",
+      "email": "jacob@wejendorp.dk",
+      "url": "https://github.com/wejendorp"
+    },
+    {
+      "name": "Jake Craige",
+      "url": "https://github.com/jakecraige"
+    },
+    {
+      "name": "Jake Marsh",
+      "url": "https://github.com/jakemmarsh"
+    },
+    {
+      "name": "Jake Mc",
+      "email": "jake.mc@icloud.com",
+      "url": "https://github.com/startswithaj"
+    },
+    {
+      "name": "Jake Verbaten",
+      "email": "raynos2@gmail.com",
+      "url": "https://github.com/Raynos"
+    },
+    {
+      "name": "Jakub Nešetřil",
+      "email": "jakub@apiary.io",
+      "url": "https://github.com/zzen"
+    },
+    {
+      "name": "James Bowes",
+      "url": "https://github.com/jbowes"
+    },
+    {
+      "name": "James Carr",
+      "email": "james.r.carr@gmail.com",
+      "url": "https://github.com/jamescarr"
+    },
+    {
+      "name": "James G. Kim",
+      "email": "jgkim@jayg.org",
+      "url": "https://github.com/jgkim"
+    },
+    {
+      "name": "James Lal",
+      "email": "james@lightsofapollo.com",
+      "url": "https://github.com/lightsofapollo"
+    },
+    {
+      "name": "James Nylen",
+      "email": "jnylen@gmail.com",
+      "url": "https://github.com/nylen"
+    },
+    {
+      "name": "Jason",
+      "url": "https://github.com/jlai"
+    },
+    {
+      "name": "Jason Barry",
+      "email": "jay@jcbarry.com",
+      "url": "https://github.com/JCBarry"
+    },
+    {
+      "name": "Javier Aranda",
+      "email": "javier.aranda.varo@gmail.com",
+      "url": "https://github.com/javierav"
+    },
+    {
+      "name": "jcreamer898",
+      "url": "https://github.com/jcreamer898"
+    },
+    {
+      "name": "Jean Ponchon",
+      "url": "https://github.com/nopnop"
+    },
+    {
+      "name": "Jeff Kunkle",
+      "url": "https://github.com/kunklejr"
+    },
+    {
+      "name": "Jeff Schilling",
+      "email": "jeff@manicwave.com",
+      "url": "https://github.com/jschilli"
+    },
+    {
+      "name": "JeongHoon Byun",
+      "email": "outsideris@gmail.com",
+      "url": "aka Outsider"
+    },
+    {
+      "name": "Jeremy Martin",
+      "url": "https://github.com/jmar777"
+    },
+    {
+      "name": "jimenglish81",
+      "url": "https://github.com/jimenglish81"
+    },
+    {
+      "name": "Jimmy Cuadra",
+      "url": "https://github.com/jimmycuadra"
+    },
+    {
+      "name": "jldailey",
+      "url": "https://github.com/jldailey"
+    },
+    {
+      "name": "jleyba",
+      "url": "https://github.com/jleyba"
+    },
+    {
+      "name": "Joey Cozza",
+      "email": "joeycozza@gmail.com",
+      "url": "https://github.com/joeycozza"
+    },
+    {
+      "name": "Johnathon Sanders",
+      "url": "https://github.com/outdooricon"
+    },
+    {
+      "name": "John Doty",
+      "email": "jrhdoty@gmail.com",
+      "url": "https://github.com/jrhdoty"
+    },
+    {
+      "name": "John Firebaugh",
+      "email": "john.firebaugh@gmail.com",
+      "url": "https://github.com/jfirebaugh"
+    },
+    {
+      "name": "John Reeves",
+      "url": "https://github.com/jonnyreeves"
+    },
+    {
+      "name": "Jo Liss",
+      "email": "joliss42@gmail.com",
+      "url": "https://github.com/joliss"
+    },
+    {
+      "name": "Jonas Dohse",
+      "url": "https://github.com/dohse"
+    },
+    {
+      "name": "Jonathan Kim",
+      "email": "hello@jkimbo.co.uk",
+      "url": "https://github.com/jkimbo"
+    },
+    {
+      "name": "Jonathan Park",
+      "email": "jonathan.daniel.park@gmail.com",
+      "url": "https://github.com/park9140"
+    },
+    {
+      "name": "jongleberry",
+      "email": "me@jongleberry.com",
+      "url": "https://github.com/jonathanong"
+    },
+    {
+      "name": "Jordan Sexton",
+      "email": "jordan@jordansexton.com",
+      "url": "https://github.com/jordansexton"
+    },
+    {
+      "name": "Joseph Spencer",
+      "url": "https://github.com/jsdevel"
+    },
+    {
+      "name": "Josh Lory",
+      "url": "https://github.com/joshlory"
+    },
+    {
+      "name": "Joshua Appelman",
+      "email": "joshua@jbna.nl",
+      "url": "https://github.com/jbnicolai"
+    },
+    {
+      "name": "Joshua Krall",
+      "email": "joshuakrall@pobox.com",
+      "url": "https://github.com/jkrall"
+    },
+    {
+      "name": "João Moreno",
+      "url": "https://github.com/joaomoreno"
+    },
+    {
+      "name": "João Paulo Bochi",
+      "email": "jpbochi@gmail.com",
+      "url": "https://github.com/jpbochi"
+    },
+    {
+      "name": "jugglinmike",
+      "url": "https://github.com/jugglinmike"
+    },
+    {
+      "name": "Julien Wajsberg",
+      "url": "https://github.com/julienw"
+    },
+    {
+      "name": "Jussi Virtanen",
+      "email": "contact@jvirtanen.org",
+      "url": "https://github.com/jvirtanen"
+    },
+    {
+      "name": "Justin DuJardin",
+      "url": "https://github.com/justindujardin"
+    },
+    {
+      "name": "Juzer Ali",
+      "email": "juzerali@live.com",
+      "url": "https://github.com/juzerali"
+    },
+    {
+      "name": "Jérémie Astori",
+      "url": "https://github.com/astorije"
+    },
+    {
+      "name": "Katie Gengler",
+      "url": "https://github.com/kategengler"
+    },
+    {
+      "name": "Kazuhito Hokamura",
+      "url": "https://github.com/hokaccha"
+    },
+    {
+      "name": "Keith Cirkel",
+      "url": "https://github.com/keithamus"
+    },
+    {
+      "name": "Kent C. Dodds",
+      "email": "kent+github@doddsfamily.us",
+      "url": "https://github.com/kentcdodds"
+    },
+    {
+      "name": "Kevin Burke",
+      "email": "kev@inburke.com",
+      "url": "https://github.com/kevinburke"
+    },
+    {
+      "name": "Kevin Conway",
+      "email": "kevinjacobconway@gmail.com",
+      "url": "https://github.com/kevinconway"
+    },
+    {
+      "name": "Kevin Kirsche",
+      "email": "Kev.Kirsche@gmail.com",
+      "url": "https://github.com/kkirsche"
+    },
+    {
+      "name": "Kirill Korolyov",
+      "email": "kirill.korolyov@gmail.com",
+      "url": "https://github.com/Dremora"
+    },
+    {
+      "name": "Koen Punt",
+      "email": "mail@koen.pt",
+      "url": "https://github.com/koenpunt"
+    },
+    {
+      "name": "Konstantin Käfer",
+      "email": "mail@kkaefer.com",
+      "url": "https://github.com/kkaefer"
+    },
+    {
+      "name": "Kris Rasmussen",
+      "url": "https://github.com/krisr"
+    },
+    {
+      "name": "Kyle Mitchell",
+      "email": "kyle@kemitchell.com",
+      "url": "https://github.com/kemitchell"
+    },
+    {
+      "name": "lakmeer",
+      "url": "https://github.com/lakmeer"
+    },
+    {
+      "name": "Liam Newman",
+      "email": "bitwiseman@gmail.com",
+      "url": "https://github.com/bitwiseman"
+    },
+    {
+      "name": "Linus Unnebäck",
+      "email": "linus@folkdatorn.se",
+      "url": "https://github.com/LinusU"
+    },
+    {
+      "name": "Long Ho",
+      "email": "holevietlong@gmail.com",
+      "url": "https://github.com/longlho"
+    },
+    {
+      "name": "László Bácsi",
+      "email": "lackac@lackac.hu",
+      "url": "https://github.com/lackac"
+    },
+    {
+      "name": "Maciej Małecki",
+      "email": "me@mmalecki.com",
+      "url": "https://github.com/mmalecki"
+    },
+    {
+      "name": "Mal Graty",
+      "url": "https://github.com/mal"
+    },
+    {
+      "name": "Marcello Bastéa-Forte",
+      "email": "marcello@cellosoft.com",
+      "url": "https://github.com/marcello3d"
+    },
+    {
+      "name": "Marc Kuo",
+      "email": "marc@routific.com",
+      "url": "https://github.com/mck-"
+    },
+    {
+      "name": "Mark Banner",
+      "url": "https://github.com/Standard8"
+    },
+    {
+      "name": "Matija Marohnić",
+      "email": "matija.marohnic@gmail.com",
+      "url": "https://github.com/silvenon"
+    },
+    {
+      "name": "Matthew Shanley",
+      "email": "matthewshanley@littlesecretsrecords.com",
+      "url": "https://github.com/arkadyan"
+    },
+    {
+      "name": "mattias-lw",
+      "url": "https://github.com/mattias-lw"
+    },
+    {
+      "name": "Matt Robenolt",
+      "email": "m@robenolt.com",
+      "url": "https://github.com/mattrobenolt"
+    },
+    {
+      "name": "Matt Smith",
+      "email": "matt@twobitfool.com",
+      "url": "https://github.com/twobitfool"
+    },
+    {
+      "name": "Max Goodman",
+      "email": "c@chromako.de",
+      "url": "https://github.com/chromakode"
+    },
+    {
+      "name": "Maximilian Antoni",
+      "email": "mail@maxantoni.de",
+      "url": "https://github.com/mantoni"
+    },
+    {
+      "name": "Merrick Christensen",
+      "email": "merrick.christensen@gmail.com",
+      "url": "https://github.com/iammerrick"
+    },
+    {
+      "name": "michael-adsk",
+      "url": "https://github.com/michael-adsk"
+    },
+    {
+      "name": "Michael Demmer",
+      "url": "https://github.com/demmer"
+    },
+    {
+      "name": "Michael Jackson",
+      "email": "mjijackson@gmail.com",
+      "url": "https://github.com/mjackson"
+    },
+    {
+      "name": "Michael Schoonmaker",
+      "email": "michael.r.schoonmaker@gmail.com",
+      "url": "https://github.com/Schoonology"
+    },
+    {
+      "name": "Michal Charemza",
+      "url": "https://github.com/michalc"
+    },
+    {
+      "name": "Mike Olson",
+      "url": "https://github.com/mwolson"
+    },
+    {
+      "name": "Mislav Marohnić",
+      "email": "mislav.marohnic@gmail.com",
+      "url": "https://github.com/mislav"
+    },
+    {
+      "name": "mrShturman",
+      "url": "https://github.com/mrShturman"
+    },
+    {
+      "name": "Nathan Alderson",
+      "email": "nathan@nathanalderson.com",
+      "url": "https://github.com/nathanalderson"
+    },
+    {
+      "name": "Nathan Black",
+      "email": "nathan@nathanblack.org",
+      "url": "https://github.com/nathanboktae"
+    },
+    {
+      "name": "Nathan Bowser",
+      "email": "nbowser@gmail.com",
+      "url": "https://github.com/nathanbowser"
+    },
+    {
+      "name": "Nathan Houle",
+      "email": "nathan+github@nathanhoule.com",
+      "url": "https://github.com/ndhoule"
+    },
+    {
+      "name": "Nathan Rajlich",
+      "email": "nathan@tootallnate.net",
+      "url": "https://github.com/TooTallNate"
+    },
+    {
+      "name": "Nick Fitzgerald",
+      "url": "https://github.com/fitzgen"
+    },
+    {
+      "name": "noirlab",
+      "url": "https://github.com/noirlab"
+    },
+    {
+      "name": "Noshir Patel",
+      "email": "nosh@blackpiano.com",
+      "url": "https://github.com/noshir-patel"
+    },
+    {
+      "name": "OlegTsyba",
+      "email": "oleg.tsyba.ua@gmail.com",
+      "url": "https://github.com/OlegTsyba"
+    },
+    {
+      "name": "omar",
+      "url": "https://github.com/omardelarosa"
+    },
+    {
+      "name": "Panu Horsmalahti",
+      "email": "panu.horsmalahti@iki.fi",
+      "url": "https://github.com/panuhorsmalahti"
+    },
+    {
+      "name": "Parker Moore",
+      "email": "email@byparker.com",
+      "url": "https://github.com/parkr"
+    },
+    {
+      "name": "Paul Armstrong",
+      "url": "https://github.com/paularmstrong"
+    },
+    {
+      "name": "Paul Miller",
+      "email": "paul+gh@paulmillr.com",
+      "url": "https://github.com/paulmillr"
+    },
+    {
+      "name": "Pavel Zubkou",
+      "url": "https://github.com/irnc"
+    },
+    {
+      "name": "Pete Hawkins",
+      "url": "https://github.com/phawk"
+    },
+    {
+      "name": "Phil Sung",
+      "email": "philbert@gmail.com",
+      "url": "https://github.com/psung"
+    },
+    {
+      "name": "Prayag Verma",
+      "email": "prayag.verma@gmail.com",
+      "url": "https://github.com/pra85"
+    },
+    {
+      "name": "qiu zuhui",
+      "email": "qiuzuhui@gmail.com",
+      "url": "https://github.com/qiuzuhui"
+    },
+    {
+      "name": "Quang Van",
+      "email": "quang@boldapps.io",
+      "url": "https://github.com/quangv"
+    },
+    {
+      "name": "Rauno",
+      "url": "https://github.com/Rauno56"
+    },
+    {
+      "name": "Refael Ackermann",
+      "email": "me@refack.com",
+      "url": "https://github.com/refack"
+    },
+    {
+      "name": "Richard Dingwall",
+      "email": "rdingwall@gmail.com",
+      "url": "https://github.com/rdingwall"
+    },
+    {
+      "name": "Richard Knop",
+      "url": "https://github.com/RichardKnop"
+    },
+    {
+      "name": "Rico Sta. Cruz",
+      "email": "hi@ricostacruz.com",
+      "url": "https://github.com/rstacruz"
+    },
+    {
+      "name": "Robert Rossmann",
+      "url": "https://github.com/Alaneor"
+    },
+    {
+      "name": "Rob Wu",
+      "email": "rob@robwu.nl",
+      "url": "https://github.com/Rob--W"
+    },
+    {
+      "name": "Romain",
+      "url": "https://github.com/rprieto"
+    },
+    {
+      "name": "Roman Neuhauser",
+      "email": "rneuhauser@sigpipe.cz",
+      "url": "https://github.com/roman-neuhauser"
+    },
+    {
+      "name": "Roman Shtylman",
+      "url": "https://github.com/defunctzombie"
+    },
+    {
+      "name": "Russ Bradberry",
+      "email": "devdazed@me.com",
+      "url": "https://github.com/devdazed"
+    },
+    {
+      "name": "Russell Munson",
+      "url": "https://github.com/rmunson"
+    },
+    {
+      "name": "Ryan",
+      "email": "ryan.shaw@min.vc",
+      "url": "https://github.com/ryan-shaw"
+    },
+    {
+      "name": "Ryan Hubbard",
+      "url": "https://github.com/ryedog"
+    },
+    {
+      "name": "Ryunosuke Sato",
+      "email": "tricknotes.rs@gmail.com",
+      "url": "https://github.com/tricknotes"
+    },
+    {
+      "name": "ryym",
+      "url": "https://github.com/ryym"
+    },
+    {
+      "name": "Salehen Shovon Rahman",
+      "email": "sal@linux.com",
+      "url": "https://github.com/shovon"
+    },
+    {
+      "name": "Salvador de la Puente González",
+      "email": "salva@unoyunodiez.com",
+      "url": "https://github.com/delapuente"
+    },
+    {
+      "name": "Sam Mussell",
+      "url": "https://github.com/smussell"
+    },
+    {
+      "name": "Samuel Goldszmidt",
+      "email": "samuel.goldszmidt@gmail.com",
+      "url": "https://github.com/ouhouhsami"
+    },
+    {
+      "name": "Sasha Koss",
+      "email": "kossnocorp@gmail.com",
+      "url": "https://github.com/kossnocorp"
+    },
+    {
+      "name": "Scott Santucci",
+      "email": "ScottFreeCode@gmail.com",
+      "url": "https://github.com/ScottFreeCode"
+    },
+    {
+      "name": "Sean Lang",
+      "email": "slang800@gmail.com",
+      "url": "https://github.com/slang800"
+    },
+    {
+      "name": "seb vincent",
+      "email": "seb.vincent@gmail.com",
+      "url": "https://github.com/sebv"
+    },
+    {
+      "name": "Seiya Konno",
+      "email": "nulltask@gmail.com",
+      "url": "https://github.com/nulltask"
+    },
+    {
+      "name": "Sergey Simonchik",
+      "url": "https://github.com/segrey"
+    },
+    {
+      "name": "Sergio Santoro",
+      "url": "https://github.com/taueres"
+    },
+    {
+      "name": "Shahar Soel",
+      "url": "https://github.com/bd82"
+    },
+    {
+      "name": "Shaine Hatch",
+      "url": "https://github.com/shaine"
+    },
+    {
+      "name": "Shiwei Wang",
+      "url": "https://github.com/wsw0108"
+    },
+    {
+      "name": "Simon Gaeremynck",
+      "url": "https://github.com/simong"
+    },
+    {
+      "name": "Simon Goumaz",
+      "url": "https://github.com/sgoumaz"
+    },
+    {
+      "name": "Sindre Sorhus",
+      "email": "sindresorhus@gmail.com",
+      "url": "https://github.com/sindresorhus"
+    },
+    {
+      "name": "slientcloud",
+      "email": "rjmuqiang@gmail.com",
+      "url": "https://github.com/silentcloud"
+    },
+    {
+      "name": "Sorin Iclanzan",
+      "url": "https://github.com/iclanzan"
+    },
+    {
+      "name": "Standa Opichal",
+      "email": "opichals@gmail.com",
+      "url": "https://github.com/opichals"
+    },
+    {
+      "name": "Stephen Mathieson",
+      "email": "me@stephenmathieson.com",
+      "url": "https://github.com/stephenmathieson"
+    },
+    {
+      "name": "Steve Mason",
+      "url": "https://github.com/spmason"
+    },
+    {
+      "name": "Stewart Taylor",
+      "email": "stewart@taylore.net",
+      "url": "https://github.com/Stewart-Taylor"
+    },
+    {
+      "name": "Sune Simonsen",
+      "email": "sune@we-knowhow.dk",
+      "url": "https://github.com/sunesimonsen"
+    },
+    {
+      "name": "Sylvain Faucherand",
+      "url": "https://github.com/slyg"
+    },
+    {
+      "name": "Takuya Nishigori",
+      "email": "nishigori.tak@gmail.com",
+      "url": "https://github.com/nishigori"
+    },
+    {
+      "name": "Taylor Gautier",
+      "url": "https://github.com/tsgautier"
+    },
+    {
+      "name": "Teddy Zeenny",
+      "url": "https://github.com/teddyzeenny"
+    },
+    {
+      "name": "Thomas Grainger",
+      "email": "https//@graingert.co.uk",
+      "url": "https://github.com/graingert"
+    },
+    {
+      "name": "Tim Ehat",
+      "url": "https://github.com/timehat"
+    },
+    {
+      "name": "Timothy Gu",
+      "email": "timothygu99@gmail.com",
+      "url": "https://github.com/TimothyGu"
+    },
+    {
+      "name": "Timo Tijhof",
+      "email": "krinklemail@gmail.com",
+      "url": "https://github.com/Krinkle"
+    },
+    {
+      "name": "Tingan Ho",
+      "email": "tingan87@gmail.com",
+      "url": "https://github.com/tinganho"
+    },
+    {
+      "name": "TJ Holowaychuk",
+      "email": "tj@vision-media.ca",
+      "url": "https://github.com/tj"
+    },
+    {
+      "name": "Tobias Bieniek",
+      "email": "tobias.bieniek@gmail.com",
+      "url": "https://github.com/Turbo87"
+    },
+    {
+      "name": "Toby Ho",
+      "email": "airportyh@gmail.com",
+      "url": "https://github.com/airportyh"
+    },
+    {
+      "name": "Todd Agulnick",
+      "url": "https://github.com/tawdle"
+    },
+    {
+      "name": "Tom Hughes",
+      "url": "https://github.com/tomhughes"
+    },
+    {
+      "name": "Tommy Montgomery",
+      "url": "https://github.com/tmont"
+    },
+    {
+      "name": "traleig1",
+      "url": "https://github.com/traleig1"
+    },
+    {
+      "name": "Travis Jeffery",
+      "email": "tj@travisjeffery.com",
+      "url": "https://github.com/travisjeffery"
+    },
+    {
+      "name": "Tyson Tate",
+      "email": "tyson@tysontate.com",
+      "url": "https://github.com/tysontate"
+    },
+    {
+      "name": "Valentin Agachi",
+      "url": "https://github.com/avaly"
+    },
+    {
+      "name": "Victor Costan",
+      "email": "victor@costan.us",
+      "url": "https://github.com/pwnall"
+    },
+    {
+      "name": "Vladimir Chernis",
+      "url": "https://github.com/vlazzle"
+    },
+    {
+      "name": "Vlad Magdalin",
+      "email": "vlad@webflow.com",
+      "url": "https://github.com/callmevlad"
+    },
+    {
+      "name": "Will Langstroth",
+      "url": "https://github.com/wlangstroth"
+    },
+    {
+      "name": "Wil Moore III",
+      "email": "wil.moore@wilmoore.com",
+      "url": "https://github.com/wilmoore"
+    },
+    {
+      "name": "Xavier Antoviaque",
+      "email": "xavier@opencraft.com",
+      "url": "https://github.com/antoviaque"
+    },
+    {
+      "name": "Xavier Damman",
+      "email": "xdamman@gmail.com",
+      "url": "https://github.com/xdamman"
+    },
+    {
+      "name": "Yanis Wang",
+      "email": "yanis.wang@gmail.com",
+      "url": "https://github.com/yaniswang"
+    },
+    {
+      "name": "yuitest",
+      "email": "developer.yuitest+github@cjhat.net",
+      "url": "https://github.com/yuitest"
+    },
+    {
+      "name": "Zhiye Li",
+      "email": "github@zhiye.li",
+      "url": "https://github.com/zhiyelee"
+    },
+    {
+      "name": "Zhouxuan Yang",
+      "email": "fool2fish@gmail.com",
+      "url": "https://github.com/fool2fish"
+    },
+    {
+      "name": "Zsolt Takács",
+      "email": "firstname at lastname dot cc",
+      "url": "https://github.com/oker1"
+    }
+  ],
+  "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git://github.com/visionmedia/mocha.git"
+    "url": "git+https://github.com/mochajs/mocha.git"
   },
-  "main": "./index",
   "bin": {
     "mocha": "./bin/mocha",
     "_mocha": "./bin/_mocha"
   },
   "engines": {
-    "node": ">= 0.4.x"
+    "node": ">= 0.10.x",
+    "npm": ">= 1.4.x"
   },
   "scripts": {
     "test": "make test"
   },
   "dependencies": {
-    "commander": "0.6.1",
-    "growl": "1.5.x",
-    "jade": "0.26.3",
-    "diff": "1.0.2",
-    "debug": "*"
+    "browser-stdout": "1.3.0",
+    "commander": "2.9.0",
+    "debug": "2.2.0",
+    "diff": "1.4.0",
+    "escape-string-regexp": "1.0.5",
+    "glob": "7.0.5",
+    "growl": "1.9.2",
+    "json3": "3.3.2",
+    "lodash.create": "3.1.1",
+    "mkdirp": "0.5.1",
+    "supports-color": "3.1.2"
   },
   "devDependencies": {
-    "should": "*",
-    "coffee-script": "1.2"
+    "assert": "^1.4.1",
+    "browserify": "^13.0.0",
+    "coffee-script": "^1.10.0",
+    "eslint": "^2.13.1",
+    "eslint-config-semistandard": "^6.0.2",
+    "eslint-config-standard": "^5.0.0",
+    "eslint-plugin-promise": "^2.0.1",
+    "eslint-plugin-standard": "1.3.2",
+    "expect.js": "^0.3.1",
+    "karma": "^1.1.0",
+    "karma-browserify": "^5.0.5",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-expect": "^1.1.2",
+    "karma-mocha": "^1.3.0",
+    "karma-phantomjs-launcher": "^0.2.3",
+    "karma-sauce-launcher": "^1.0.0",
+    "karma-spec-reporter": "0.0.26",
+    "os-name": "^2.0.1",
+    "phantomjs": "1.9.8",
+    "rimraf": "^2.5.2",
+    "should": "^9.0.2",
+    "through2": "^2.0.1",
+    "watchify": "^3.7.0"
   },
+  "files": [
+    "bin",
+    "images",
+    "lib",
+    "index.js",
+    "mocha.css",
+    "mocha.js",
+    "browser-entry.js",
+    "LICENSE",
+    "bower.json"
+  ],
+  "browser": {
+    "debug": "./lib/browser/debug.js",
+    "events": "./lib/browser/events.js",
+    "tty": "./lib/browser/tty.js",
+    "./index.js": "./browser-entry.js",
+    "fs": false,
+    "glob": false,
+    "path": false,
+    "supports-color": false
+  },
+  "homepage": "https://mochajs.org",
+  "logo": "https://cldup.com/S9uQ-cOLYz.svg",
+  "gitHead": "b51e36014d9c6db07aee3057579c35315ec4efd7",
+  "bugs": {
+    "url": "https://github.com/mochajs/mocha/issues"
+  },
+  "_id": "mocha@3.2.0",
+  "_shasum": "7dc4f45e5088075171a68896814e6ae9eb7a85e3",
+  "_from": "mocha@latest",
+  "_npmVersion": "3.10.9",
+  "_nodeVersion": "7.2.0",
   "_npmUser": {
-    "name": "mrege",
-    "email": "matias.rege@mercadolibre.com"
+    "name": "boneskull",
+    "email": "boneskull@boneskull.com"
   },
-  "_id": "mocha@1.2.2",
-  "optionalDependencies": {},
-  "_engineSupported": true,
-  "_npmVersion": "1.1.4",
-  "_nodeVersion": "v0.6.12",
-  "_defaultsLoaded": true,
-  "_from": "mocha@1.2.x"
+  "dist": {
+    "shasum": "7dc4f45e5088075171a68896814e6ae9eb7a85e3",
+    "tarball": "https://registry.npmjs.org/mocha/-/mocha-3.2.0.tgz"
+  },
+  "maintainers": [
+    {
+      "name": "boneskull",
+      "email": "boneskull@boneskull.com"
+    },
+    {
+      "name": "dasilvacontin",
+      "email": "dasilvacontin@gmail.com"
+    }
+  ],
+  "_npmOperationalInternal": {
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/mocha-3.2.0.tgz_1480031903482_0.1920817163772881"
+  },
+  "directories": {},
+  "_resolved": "https://registry.npmjs.org/mocha/-/mocha-3.2.0.tgz",
+  "readme": "ERROR: No README data found!"
 }
